Revert "feat: show image preview next to image input field"
This reverts commit e77ccfa6b73dfc61e33829e730017c95bd466ef5.
